FT metadata, web4, docs (#30)
Expose the ft_metadata method to the contract wasm. It is needed by wallets to properly show the balance.
Docs about setting up the Fungible Token contract
Implementation of web4_get to be able to serve a web page from the contract
Scripts for bundling for web4, using the AI proxy frontend as an example
Docs about submitting javascript code to the contract
Docs for deploying the AI proxy to Spin cloud